fix: Use `latest` dist tag as default for canary package versions
Based on the findings and the feedback in https://github.com/web-infra-dev/rspack/pull/10538, the default version for `@rspack-canary/core` was changed from `canary` dist tag to `latest` dist tag.

After merging related PR in the core project, for the `@rspack-canary/core` the `latest` dist tag should be used by default. The install script was changed accordingly. In this PR in case of `canary` version, the script will look into canary variants of the packages to find the latest version. In any other cases, the script should executes as before.